Separate shared metrics code
