+++ src/shadowbox/server/metrics.ts
@@ -14,23 +14,18 @@
 
 import * as events from 'events';
 import * as fs from 'fs';
-import * as url from 'url';
 
 import * as file_read from '../infrastructure/file_read';
-import * as follow_redirects from '../infrastructure/follow_redirects';
-import * as ip_location from '../infrastructure/ip_location';
 import * as logging from '../infrastructure/logging';
 import {AccessKeyId} from '../model/access_key';
 import {DataUsageByUser, LastHourMetricsReadyCallback, PerUserStats, Stats} from '../model/metrics';
 
-import * as ip_util from './ip_util';
-
-const MS_PER_HOUR = 60 * 60 * 1000;
+import {SharedStats} from './shared_metrics';
 
 interface PersistentStatsStoredData {
-  // Serialized TransferStats object.
+  // Serialized ManagerStats object.
   transferStats: string;
-  // Serialized ConnectionStats object.
+  // Serialized SharedStats object.
   hourlyMetrics: string;
 }
 
@@ -38,8 +33,8 @@ interface PersistentStatsStoredData {
 // a PersistentStatsStoredData object.
 export class PersistentStats implements Stats {
   private static readonly MAX_STATS_FILE_AGE_MS = 5000;
-  private transferStats: TransferStats;
-  private connectionStats: ConnectionStats;
+  private managerStats: ManagerStats;
+  private sharedStats: SharedStats;
   private dirty = false;
   private eventEmitter = new events.EventEmitter();
   private static readonly LAST_HOUR_METRICS_READY_EVENT = 'lastHourMetricsReady';
@@ -48,11 +43,11 @@ export class PersistentStats implements Stats {
     // Initialize stats from saved file, if available.
     const persistedStateObj = this.readStateFile();
     if (persistedStateObj) {
-      this.transferStats = new TransferStats(persistedStateObj.transferStats);
-      this.connectionStats = new ConnectionStats(persistedStateObj.hourlyMetrics);
+      this.managerStats = new ManagerStats(persistedStateObj.transferStats);
+      this.sharedStats = new SharedStats(persistedStateObj.hourlyMetrics);
     } else {
-      this.transferStats = new TransferStats();
-      this.connectionStats = new ConnectionStats();
+      this.managerStats = new ManagerStats();
+      this.sharedStats = new SharedStats();
     }
 
     // Set write interval.
@@ -66,15 +61,15 @@ export class PersistentStats implements Stats {
       userId: AccessKeyId, metricsUserId: AccessKeyId, numBytes: number, ipAddresses: string[]) {
     // Pass the userId (sequence number) to transferStats as this data is returned to the Outline
     // manager which relies on the userId sequence number.
-    this.transferStats.recordBytesTransferred(userId, numBytes);
+    this.managerStats.recordBytesTransferred(userId, numBytes);
     // Pass metricsUserId (uuid, rather than sequence number) to connectionStats
     // as these values may be reported to the Outline metrics server.
-    this.connectionStats.recordBytesTransferred(metricsUserId, numBytes, ipAddresses);
+    this.sharedStats.recordBytesTransferred(metricsUserId, numBytes, ipAddresses);
     this.dirty = true;
   }
 
   public get30DayByteTransfer(): DataUsageByUser {
-    return this.transferStats.get30DayByteTransfer();
+    return this.managerStats.get30DayByteTransfer();
   }
 
   public onLastHourMetricsReady(callback: LastHourMetricsReadyCallback) {
@@ -82,7 +77,7 @@ export class PersistentStats implements Stats {
 
     // Check if an hourly metrics report is already due (e.g. if server was shutdown over an
     // hour ago and just restarted).
-    if (getHoursSinceDatetime(this.connectionStats.startDatetime) >= 1) {
+    if (getHoursSinceDatetime(this.sharedStats.startDatetime) >= 1) {
       this.generateHourlyReport();
     }
   }
@@ -93,8 +88,8 @@ export class PersistentStats implements Stats {
     }
 
     const statsSerialized = JSON.stringify({
-      transferStats: this.transferStats.serialize(),
-      hourlyMetrics: this.connectionStats.serialize()
+      transferStats: this.managerStats.serialize(),
+      hourlyMetrics: this.sharedStats.serialize()
     });
 
     // Write to temporary file, then move that temporary file to the
@@ -112,18 +107,18 @@ export class PersistentStats implements Stats {
   }
 
   private generateHourlyReport(): void {
-    if (this.connectionStats.lastHourUserStats.size === 0) {
+    if (this.sharedStats.lastHourUserStats.size === 0) {
       // No connection stats to report.
       return;
     }
 
     this.eventEmitter.emit(
-        PersistentStats.LAST_HOUR_METRICS_READY_EVENT, this.connectionStats.startDatetime,
+        PersistentStats.LAST_HOUR_METRICS_READY_EVENT, this.sharedStats.startDatetime,
         new Date(),  // endDatetime is the current date and time.
-        this.connectionStats.lastHourUserStats);
+        this.sharedStats.lastHourUserStats);
 
     // Reset connection stats to begin recording the next hour.
-    this.connectionStats.reset();
+    this.sharedStats.reset();
 
     // Update hasChange so we know to persist stats.
     this.dirty = true;
@@ -142,8 +137,9 @@ export class PersistentStats implements Stats {
   }
 }
 
-// TransferStats keeps track of the number of bytes transferred per user, per day.
-class TransferStats {
+// ManagerStats keeps track of the number of bytes transferred per user, per day.
+// Surfaced by the manager service to display on the Manager UI.
+class ManagerStats {
   // Key is a string in the form "userId-dateInYYYYMMDD", e.g. "3-20170726".
   private dailyUserBytesTransferred: Map<string, number>;
   // Set of all User IDs for whom we have transfer stats.
@@ -212,121 +208,7 @@ class TransferStats {
   }
 }
 
-// Keeps track of the connection stats per user, sine the startDatetime.
-class ConnectionStats {
-  // Date+time at which we started recording connection stats, e.g.
-  // in case this object is constructed from data written to disk.
-  public startDatetime: Date;
-
-  // Map from the metrics AccessKeyId to stats (bytes transferred, IP addresses).
-  public lastHourUserStats: Map<AccessKeyId, PerUserStats>;
-
-  constructor(serializedObject?: {}) {
-    if (serializedObject) {
-      this.deserialize(serializedObject);
-    } else {
-      this.startDatetime = new Date();
-      this.lastHourUserStats = new Map();
-    }
-  }
-
-  // CONSIDER: accepting hashedIpAddresses, which can be persisted to disk
-  // and reported to the metrics server (to approximate number of devices per userId).
-  public recordBytesTransferred(userId: AccessKeyId, numBytes: number, ipAddresses: string[]) {
-    const perUserStats = this.lastHourUserStats.get(userId) ||
-        {bytesTransferred: 0, anonymizedIpAddresses: new Set<string>()};
-    perUserStats.bytesTransferred += numBytes;
-    const anonymizedIpAddresses = getAnonymizedAndDedupedIpAddresses(ipAddresses);
-    for (const ip of anonymizedIpAddresses) {
-      perUserStats.anonymizedIpAddresses.add(ip);
-    }
-    this.lastHourUserStats.set(userId, perUserStats);
-  }
-
-  public reset(): void {
-    this.lastHourUserStats = new Map<AccessKeyId, PerUserStats>();
-    this.startDatetime = new Date();
-  }
-
-  // Returns the state of this object, e.g.
-  // {"startTimestamp":1502896650353,"lastHourUserStatsObj":{"0":{"bytesTransferred":100,"anonymizedIpAddresses":["2620:0:1003:0:0:0:0:0","5.2.79.0"]}}}
-  public serialize(): {} {
-    // lastHourUserStats is a Map containing Set structures.  Convert to an object
-    // with array values.
-    const lastHourUserStatsObj = {};
-    this.lastHourUserStats.forEach((perUserStats, userId) => {
-      lastHourUserStatsObj[userId] = {
-        bytesTransferred: perUserStats.bytesTransferred,
-        anonymizedIpAddresses: [...perUserStats.anonymizedIpAddresses]
-      };
-    });
-    return {startTimestamp: this.startDatetime.getTime(), lastHourUserStatsObj};
-  }
-
-  private deserialize(serializedObject: {}) {
-    // Convert type of lastHourUserStatsObj from Object containing Arrays to
-    // Map containing Sets.
-    const lastHourUserStatsMap = new Map<AccessKeyId, PerUserStats>();
-    Object.keys(serializedObject['lastHourUserStatsObj']).map((userId) => {
-      const perUserStatsObj = serializedObject['lastHourUserStatsObj'][userId];
-      lastHourUserStatsMap.set(userId, {
-        bytesTransferred: perUserStatsObj.bytesTransferred,
-        anonymizedIpAddresses: new Set(perUserStatsObj.anonymizedIpAddresses)
-      });
-    });
-
-    this.startDatetime = new Date(serializedObject['startTimestamp']);
-    this.lastHourUserStats = lastHourUserStatsMap;
-  }
-}
-
-export function getHourlyServerMetricsReport(
-    serverId: string, startDatetime: Date, endDatetime: Date,
-    lastHourUserStats: Map<AccessKeyId, PerUserStats>,
-    ipLocationService: ip_location.IpLocationService): Promise<HourlyServerMetricsReport|null> {
-  if (lastHourUserStats.size === 0) {
-    // Stats are empty, no need to post a report
-    return Promise.resolve(null);
-  }
-  // convert lastHourUserStats to an array HourlyUserMetricsReport
-  const userReportPromises = [];
-  lastHourUserStats.forEach((perUserStats, userId) => {
-    userReportPromises.push(getHourlyUserMetricsReport(userId, perUserStats, ipLocationService));
-  });
-  return Promise.all(userReportPromises).then((userReports: HourlyUserMetricsReport[]) => {
-    // Remove any userReports containing sanctioned countries, and return
-    // null if no reports remain with un-sanctioned countries.
-    userReports = getWithoutSanctionedReports(userReports);
-    if (userReports.length === 0) {
-      return null;
-    }
-    return {
-      serverId,
-      startUtcMs: startDatetime.getTime(),
-      endUtcMs: endDatetime.getTime(),
-      userReports
-    };
-  });
-}
-
-export function postHourlyServerMetricsReports(
-    report: HourlyServerMetricsReport, metricsUrl: string) {
-  const options = {
-    url: metricsUrl,
-    headers: {'Content-Type': 'application/json'},
-    method: 'POST',
-    body: JSON.stringify(report)
-  };
-  logging.info('Posting metrics: ' + JSON.stringify(options));
-  return follow_redirects.requestFollowRedirectsWithSameMethodAndBody(
-      options, (error, response, body) => {
-        if (error) {
-          logging.error(`Error posting metrics: ${error}`);
-          return;
-        }
-        logging.info('Metrics server responded with status ' + response.statusCode);
-      });
-}
+const MS_PER_HOUR = 60 * 60 * 1000;
 
 function setHourlyInterval(callback: Function) {
   const msUntilNextHour = MS_PER_HOUR - (Date.now() % MS_PER_HOUR);
@@ -340,69 +222,4 @@ function setHourlyInterval(callback: Function) {
 function getHoursSinceDatetime(d: Date): number {
   const deltaMs = Date.now() - d.getTime();
   return deltaMs / (MS_PER_HOUR);
-}
-
-interface HourlyServerMetricsReport {
-  serverId: string;
-  startUtcMs: number;
-  endUtcMs: number;
-  userReports: HourlyUserMetricsReport[];
-}
-
-interface HourlyUserMetricsReport {
-  userId: string;
-  countries: string[];
-  bytesTransferred: number;
-}
-
-function getHourlyUserMetricsReport(
-    userId: AccessKeyId, perUserStats: PerUserStats,
-    ipLocationService: ip_location.IpLocationService): Promise<HourlyUserMetricsReport> {
-  const countryPromises = [];
-  for (const ip of perUserStats.anonymizedIpAddresses) {
-    const countryPromise = ipLocationService.countryForIp(ip).catch((e) => {
-      logging.warn(`Failed countryForIp call: ${e}`);
-      return 'ERROR';
-    });
-    countryPromises.push(countryPromise);
-  }
-  return Promise.all(countryPromises).then((countries: string[]) => {
-    return {
-      userId,
-      bytesTransferred: perUserStats.bytesTransferred,
-      countries: getWithoutDuplicates(countries)
-    };
-  });
-}
-
-function getAnonymizedAndDedupedIpAddresses(ipAddresses: string[]): Set<string> {
-  const s = new Set<string>();
-  for (const ip of ipAddresses) {
-    try {
-      s.add(ip_util.anonymizeIp(ip));
-    } catch (err) {
-      logging.error('error anonymizing IP address: ' + ip + ', ' + err);
-    }
-  }
-  return s;
-}
-
-// Return an array with the duplicate elements removed.
-function getWithoutDuplicates<T>(a: T[]): T[] {
-  return [...new Set(a)];
-}
-
-function getWithoutSanctionedReports(userReports: HourlyUserMetricsReport[]):
-    HourlyUserMetricsReport[] {
-  const sanctionedCountries = ['CU', 'IR', 'KP', 'SY'];
-  const filteredReports = [];
-  for (const userReport of userReports) {
-    userReport.countries = userReport.countries.filter((country) => {
-      return sanctionedCountries.indexOf(country) === -1;
-    });
-    if (userReport.countries.length > 0) {
-      filteredReports.push(userReport);
-    }
-  }
-  return filteredReports;
 }
